UI Update
moved to WPF, added UI, table selector with openfiledialog, text previewer, copy hex to clipboard and message boxes.
