+++ DialoguePreview.xaml
@@ -1,0 +1,32 @@
+﻿<Window x:Class="HedgeHex.DialoguePreview"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:HedgeHex"
+        mc:Ignorable="d"
+        Title="Text Preview" Height="450" Width="800" Foreground="{x:Null}" ResizeMode="CanMinimize" Closing="Window_Closing">
+    <Window.Resources>
+        <FontFamily x:Key="font" >/fonts/#FOT-Seurat Pro M</FontFamily>
+    </Window.Resources>
+    <Window.Background>
+        <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
+            <GradientStop Color="#FFB74343"/>
+            <GradientStop Color="#FFEACA39" Offset="1"/>
+        </LinearGradientBrush>
+    </Window.Background>
+    <Grid>
+
+        <TextBlock x:Name="Text1" FontSize="20" HorizontalAlignment="Center" 
+               FontFamily="{StaticResource font}" Width="758" Foreground="White" Panel.ZIndex="1" VerticalAlignment="Center" Height="386">TEST
+            <TextBlock.Effect>
+                <DropShadowEffect
+      ShadowDepth="1.7"
+      Direction="319.7"
+      Color="Black"
+      Opacity="1"
+     BlurRadius="0" RenderingBias="Performance"/>
+            </TextBlock.Effect>
+        </TextBlock>
+    </Grid>
+</Window>
