+++ HedgeHex.csproj
@@ -1,14 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>net6.0-windows</TargetFramework>
     <Nullable>enable</Nullable>
+    <UseWPF>true</UseWPF>
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="Program.cs~RF4f50b.TMP" />
+    <None Remove="fonts\FOT-Seurat Pro M.otf" />
   </ItemGroup>
 
   <ItemGroup>
@@ -16,7 +16,20 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="tables\common.json">
+    <Folder Include="tables\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Resource Include="fonts\FOT-Seurat Pro M.otf">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Resource>
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="tables\Common.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="tables\In-Stage.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
