--- HedgeHex.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32602.215
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HedgeHex", "HedgeHex.csproj", "{75511DEF-1DED-4F21-898D-8E607F3759B8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{75511DEF-1DED-4F21-898D-8E607F3759B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75511DEF-1DED-4F21-898D-8E607F3759B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{75511DEF-1DED-4F21-898D-8E607F3759B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{75511DEF-1DED-4F21-898D-8E607F3759B8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {752845A0-E40B-447E-83D7-D3B6F3E1C1B3}
-	EndGlobalSection
-EndGlobal
