+++ MainWindow.xaml
@@ -1,0 +1,18 @@
+﻿<Window x:Class="HedgeHex.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:HedgeHex"
+        mc:Ignorable="d"
+        Title="HedgeHex" Height="359" Width="586" WindowStyle="SingleBorderWindow" Background="#FF1F1F1F" Opacity="1" ResizeMode="CanMinimize">
+    <Grid Background="#001F1F1F">
+        <TextBox x:Name="Textbox" HorizontalAlignment="Left" Margin="9,44,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="552" Background="#FF484848" Foreground="White" BorderBrush="#00ABADB3" Height="36" FontSize="25" FontFamily="Microsoft YaHei UI Light"/>
+        <Button x:Name="Convert" Content="Convert" HorizontalAlignment="Left" Margin="9,165,0,0" VerticalAlignment="Top" Height="59" Width="552" Background="#FF3C7737" Foreground="White" BorderBrush="#00707070" FontSize="30" Click="Convert_Click" FontFamily="Segoe UI"/>
+        <Button x:Name="TranslationTableButton" Content="Select Translation Table" HorizontalAlignment="Left" VerticalAlignment="Top" Width="552" Height="36" FontFamily="Microsoft YaHei UI Light" FontSize="22" Background="#FF484848" BorderBrush="#00707070" Foreground="White" Click="TranslationTableButton_Click" Margin="9,115,0,0"/>
+        <TextBlock x:Name="ConvertTextLabel" HorizontalAlignment="Left" Margin="9,9,0,0" TextWrapping="Wrap" Text="Text that will be converted:" VerticalAlignment="Top" Height="30" Width="552" Foreground="White" FontSize="20" FontFamily="Microsoft YaHei UI Light"/>
+        <TextBlock x:Name="ConvertTextLabel_Copy" HorizontalAlignment="Left" Margin="9,80,0,0" TextWrapping="Wrap" Text="Translation Table that will be used (.json)" VerticalAlignment="Top" Height="30" Width="552" Foreground="White" FontSize="20" FontFamily="Microsoft YaHei UI Light"/>
+        <Button x:Name="Import_Text" Content="Import" HorizontalAlignment="Left" Margin="469,19,0,0" VerticalAlignment="Top" Width="92" Background="DimGray" BorderBrush="#FF2F2F2F" Foreground="White" Click="Import_Text_Click" FontFamily="Microsoft YaHei UI Light"/>
+        <Button x:Name="PreviewButton" Content="Preview Text" HorizontalAlignment="Left" Margin="10,240,0,0" VerticalAlignment="Top" Height="59" Width="551" Background="#FFFF9144" Foreground="White" BorderBrush="#00707070" FontSize="30" FontFamily="Segoe UI" Click="PreviewButton_Click"/>
+    </Grid>
+</Window>
