--- README.md
@@ -1,4 +1,0 @@
-# HedgeHex
-A tool to translate text into Unleashed FCO-compatible HEX
-
-Findings done by Hedgeturd
